+++ CHANGELOG.md
@@ -338,4 +338,5 @@
 * Added Brightness slider to options panel
 * Quality level is now stored in ConfigState and applied at runtime
 * CoreParams.DefaultJsonSerializerSettings is now used in a few more places
+* Added GetDescriptorForAxis and GetDescriptorForButton APIs to input mappers for getting names of mappings etc [Experimental]
 * Implemented migration for save games on load [Untested]